Changes as per feedback
